refactor(acp): move parent ACP client into contrib plugin
- Relocate executor, registry, and entry pins out of adapters
- Keep wrap_docker_exec and layer-C writer on Core
- Point inventory, lock, and pin-parity at the plugin package
