+++ src/bora/plugins/contrib/acp/__init__.py
@@ -1,13 +1,18 @@
-"""ACP first-party contrib: provide(executor) + image/trajectory hooks (Spec 01).
+"""ACP first-party contrib: parent client, entry registry, executor SPI.
 
-Not a full external ACP plugin package. Wrappers live here; protocol client code
-may remain under ``bora.adapters.acp`` and is imported by the factory.
+Not an external ``plugins/acp`` package and not installed via ``bora plugin
+install``. Official entries stay bake-in on ``docker/attempt``.
 """
 
 from __future__ import annotations
 
 from typing import Any
 
+from bora.plugins.contrib.acp.executor import AcpExecutor
+from bora.plugins.contrib.acp.resolve import resolve_acp_executor
+from bora.plugins.contrib.acp.trajectory_map import acp_session_events_to_bora
+from bora.plugins.contrib.acp.types import ProcessLauncher
+from bora.plugins.contrib.acp.usage import normalize_acp_usage
 from bora.plugins.protocol import ExecutorSPI
 from bora.plugins.registry import ExtensionRegistry
 from bora.plugins.slots import (
@@ -21,7 +26,7 @@ ACP_PRIORITY = 100
 
 
 class AcpExecutorSPI(ExecutorSPI):
-    """ExecutorSPI facade over adapters.acp.AcpExecutor."""
+    """ExecutorSPI facade over the in-plugin AcpExecutor."""
 
     kind = "acp"
 
@@ -47,7 +52,7 @@ class AcpExecutorSPI(ExecutorSPI):
                 "acp_entry_required",
                 kind="extension_materialize_failed",
             )
-        from bora.adapters.acp import AcpExecutor
+        from bora.plugins.contrib.acp.executor import AcpExecutor
 
         self.profile_id = profile_id
         self._entry_id = str(entry).strip()
@@ -92,9 +97,9 @@ class AcpExecutorSPI(ExecutorSPI):
 
     def bind_to_target(self, placement: Any) -> AcpExecutorSPI:
         """Attach parent ACP client to the Attempt container via docker exec."""
-        from bora.adapters.acp import AcpExecutor
         from bora.adapters.agent_container import wrap_docker_exec
         from bora.adapters.child_env import cli_env_for_container
+        from bora.plugins.contrib.acp.executor import AcpExecutor
 
         child_env = cli_env_for_container(
             self._entry_id, api_key_env=self._api_key_env, base_url=self._base_url
@@ -207,4 +212,13 @@ def register_acp_contrib(registry: ExtensionRegistry) -> None:
     )
 
 
-__all__ = ["PLUGIN_ID", "AcpExecutorSPI", "register_acp_contrib"]
+__all__ = [
+    "PLUGIN_ID",
+    "AcpExecutor",
+    "AcpExecutorSPI",
+    "ProcessLauncher",
+    "acp_session_events_to_bora",
+    "normalize_acp_usage",
+    "register_acp_contrib",
+    "resolve_acp_executor",
+]
