+++ src/bora/plugins/contrib/acp/acp_entries.json
